multi-agent-ai-system/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
langchain
langgraph
langchain-groq
langchain-community
duckduckgo-search
groq
pypdf
pdfplumber
huey
aio-pika
psycopg2-binary
redis
deprecated
wrapt
docker
langgraph-checkpoint-sqlite